Change model for Links table
